JD03_03 DAO + Abstract
